Implement lisp REPL mode
This adds a true REPL mode for entering lisp sexpressions with support
for all the usual Julia REPL features. Uses `lisp> ` as the prompt text,
prompt color is purple and the mode is set to `)`. Any changes to those
defaults you'd rather have?

Multiline text entry is supported by checking whether the text is a
complete sexpr with `Lisp.read`. Is there a better/cheaper way to
check that the buffer is a valid sexpr?

Also adds the lisp mode at module init time so that users don't have to
call `initrepl` manually. This does mean that the prompt, etc. can't
be customised easily. Would you rather leave it to the user to setup
themselves?